requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.6.3
motor==3.3.2
beanie==1.26.0
passlib[bcrypt]==1.7.4
openai>=1.35.0
playwright==1.45.0
python-multipart==0.0.9
pydantic-settings>=2.3.0
httpx>=0.23.0,<0.28.0
playwright-stealth>=2.0.0
